# add byte comparisons