--- a/.gitignore
+++ b/.gitignore
@@ -7,3 +7,4 @@ dist/*
 sha256sum.mojo_csv.mojopkg.txt
 mojo_csv.mojopkg
 output
+tests/datablist/products-2000000.csv