--- a/bench_threaded.mojo
+++ b/bench_threaded.mojo
@@ -27,7 +27,9 @@ fn bench_multi_threaded_medium() capturing:
 
 fn bench_single_threaded_small() capturing:
     try:
-        var in_csv: Path = cwd().joinpath("tests/datablist/organizations-1000.csv")
+        var in_csv: Path = cwd().joinpath(
+            "tests/datablist/organizations-1000.csv"
+        )
         var _ = CsvReader(in_csv)
     except:
         print("error in single threaded small")
@@ -36,10 +38,34 @@ fn bench_single_threaded_small() capturing:
 
 fn bench_multi_threaded_small() capturing:
     try:
-        var in_csv: Path = cwd().joinpath("tests/datablist/organizations-1000.csv")
+        var in_csv: Path = cwd().joinpath(
+            "tests/datablist/organizations-1000.csv"
+        )
         var _ = ThreadedCsvReader(in_csv)
     except:
         print("error in multi threaded small")
+        exit()
+
+
+fn bench_single_threaded_large() capturing:
+    try:
+        var in_csv: Path = cwd().joinpath(
+            "tests/datablist/products-2000000.csv"
+        )
+        var _ = CsvReader(in_csv)
+    except:
+        print("error in single threaded large")
+        exit()
+
+
+fn bench_multi_threaded_large() capturing:
+    try:
+        var in_csv: Path = cwd().joinpath(
+            "tests/datablist/products-2000000.csv"
+        )
+        var _ = ThreadedCsvReader(in_csv)
+    except:
+        print("error in multi threaded large")
         exit()
 
 
@@ -91,6 +117,29 @@ fn main():
     print("Speedup:", round(speedup_medium, 2), "x")
     print("-------------------------")
 
+    # Test large file (2M rows)
+    print("Large file benchmark (2,000,000 rows):")
+    print("Single-threaded:")
+    var time_single_large: Float64 = 0
+    for _ in range(3):  # Fewer iterations for large file
+        var elapsed = time_function[bench_single_threaded_large]()
+        time_single_large += elapsed / 1000000
+    var avg_single_large = time_single_large / 3
+    print("Average time:", round(avg_single_large, 6), "ms")
+
+    print("Multi-threaded:")
+    var time_multi_large: Float64 = 0
+    for _ in range(3):  # Fewer iterations for large file
+        var elapsed = time_function[bench_multi_threaded_large]()
+        time_multi_large += elapsed / 1000000
+    var avg_multi_large = time_multi_large / 3
+    print("Average time:", round(avg_multi_large, 6), "ms")
+
+    var speedup_large = avg_single_large / avg_multi_large
+    print("Speedup:", round(speedup_large, 2), "x")
+    print("-------------------------")
+
     print("Summary:")
     print("Small file speedup:", round(speedup_small, 2), "x")
     print("Medium file speedup:", round(speedup_medium, 2), "x")
+    print("Large file speedup:", round(speedup_large, 2), "x")