- Add byte-level comparison optimization
- Use logical cores to determine thread count for better compatibility
- Added Large File benchmark
- Single Threaded Performance Improvements, ~3x speedup
- Threaded Performance Improvements, +100% over previous / ~+40% over single threaded on large files

average time in ms for micro file:
0.0076ms -> .0018ms
-------------------------
average time in ms for mini file:
0.24ms -> .073ms
-------------------------
average time in ms for small file:
1.39ms -> .41ms -> [T] .39ms
-------------------------
running benchmark for medium csv:
average time in ms for medium file:
121.22ms -> 36.46ms -> [T] 24.84ms
-------------------------
running benchmark for large csv:
average time in ms for large file:
3582.88ms -> 1253.20ms -> [T] 845.33ms
-------------------------